fix(multiple): Reworked the way conan is started through supervisord
I've noticed that stopping conan through supervisord does nothing.
After a few searches I found out it's related with how the previously used xvfb-run tool works.
It will just fork a Xvfb process that is detached from xvbd-run.

Per nature supervisord uses signals sent to the process it started to stop it. It will then not work in that case.
That why I reworked the things so that supervisord directly call wine64 to open the Conan Exiles executable.
So when we ask supervisord to stop the process it will kill TERM this one and hopefully really stop it.

Also now Xvfb is started once for all and sent to background at container boot and not from supervisord anymore.
I also added checks in the controller bash script to check that server has been successfully stopped (for server_stop) and that it's not already running (for server_start).

I hope these changes will work around what can go wrong when an update is available :)